moved 'auth' out of 'api' folder into 'agentkernel' folder
