+++ ak-py/src/agentkernel/api/__init__.py
@@ -11,11 +11,5 @@ try:
 except importlib.metadata.PackageNotFoundError:
     __version__ = "0.1.0"
 
-from .auth import AuthValidator, ValidationContext, ValidationResult
-
-try:
-    # AuthValidator does not need these, these need fastapi and other libraries which are not available in AuthValidator as it doesn't need them
-    from .handler import AgentRESTRequestHandler, RESTRequestHandler
-    from .http import RESTAPI
-except ImportError:
-    pass
+from .handler import AgentRESTRequestHandler, RESTRequestHandler
+from .http import RESTAPI
