+++ examples/aws-containerized/crewai/app.py
@@ -1,7 +1,8 @@
 import os
 from typing import Optional
 
-from agentkernel.api import RESTAPI, AgentRESTRequestHandler, AuthValidator, ValidationContext, ValidationResult
+from agentkernel.api import RESTAPI, AgentRESTRequestHandler
+from agentkernel.auth import AuthValidator, ValidationContext, ValidationResult
 from agentkernel.crewai import CrewAIModule
 from fastapi import APIRouter
